// apps/web/src/types/dao.ts
export type CosmosMsgFor_Empty =
  | {
      bank: BankMsg
    }
  | {
      custom: Empty
    }
  | {
      staking: StakingMsg
    }
  | {
      distribution: DistributionMsg
    }
  | {
      ibc: IbcMsg
    }
  | {
      wasm: WasmMsg
    }
  | {
      gov: GovMsg
    }
  | StargateMsg
// ...
export enum DepositRefundPolicy {
  Always = 'always',
  OnlyPassed = 'only_passed',
  Never = 'never',
}

export type ExecuteMsg =
  | {
      propose: {
        msg: ProposeMessage
        [k: string]: unknown
      }
    }
  | {
      update_config:
        | {
            allow_revoting: boolean
            close_proposal_on_execution_failure: boolean
            dao: string
            max_voting_period: Duration
            min_voting_period?: Duration | null
            only_members_execute: boolean
            threshold: Threshold
            [k: string]: unknown
          }
        | {
            deposit_info?: UncheckedDepositInfo | null
            open_proposal_submission: boolean
            [k: string]: unknown
          }
    }
  | {
      withdraw: {
        denom?: UncheckedDenom | null
        [k: string]: unknown
      }
    }
  | {
      extension: {
        msg: Empty
        [k: string]: unknown
      }
    }
  | {
      proposal_created_hook: {
        proposal_id: number
        proposer: string
        [k: string]: unknown
      }
    }
  | {
      proposal_completed_hook: {
        new_status: Status
        proposal_id: number
        [k: string]: unknown
      }
    }
export type ProposeMessage = {
  propose: {
    description: string
    msgs: CosmosMsgFor_Empty[]
    title: string
  }
}

export interface UncheckedDepositInfo {
  amount: Uint128
  denom: DepositToken
  refund_policy: DepositRefundPolicy
  [k: string]: unknown
}
export type DepositToken =
  | {
      token: {
        denom: UncheckedDenom
      }
    }
  | {
      voting_module_token: Empty
    }
export type UncheckedDenom =
  | {
      native: string
    }
  | {
      cw20: string
    }
export type Status = 'open' | 'rejected' | 'passed' | 'executed' | 'closed' | 'execution_failed'

export type Duration =
  | {
      height: number
    }
  | {
      time: number
    }
export type Threshold =
  | {
      absolute_percentage: {
        percentage: PercentageThreshold
      }
    }
  | {
      threshold_quorum: {
        quorum: PercentageThreshold
        threshold: PercentageThreshold
      }
    }
  | {
      absolute_count: {
        threshold: Uint128
      }
    }
export type PercentageThreshold =
  | {
      majority: Empty
    }
  | {
      percent: Decimal
    }